Removed BigDecimal, Add getRatingEndpoint

+++ src/main/java/com/github/flooooooooooorian/meinkochbuch/dtos/ingredient/IngredientCreationDto.java
@@ -6,8 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -17,5 +15,5 @@ public class IngredientCreationDto {
     private String text;
 
     private BaseIngredient baseIngredient;
-    private BigDecimal amount;
+    private double amount;
 }
